+++ src/block-types/header/index.js
@@ -18,7 +18,7 @@ import { registerBlockType, createBlock } from "@wordpress/blocks";
  *
  * @see https://www.npmjs.com/package/@wordpress/scripts#using-css
  */
-// import "./style.scss";
+import "./style.scss";
 
 /**
  * Styles applied only in the editor.